docs(chromecast): Add docs on iframe embed limitations
We show an arbitrary URL on the device by embedding it into an iframe
in our Chromecast receiver app.  However, sites can prevent
iframe-embedding with the X-Frame-Options header.

Though this should not be an issue for a test runner, this may affect
other URLs.  Unfortunately, there is no way for the receiver app to
detect when this has happened.  See also #8.

This adds docs to explain this limitation, and updates our CLI example
to something that will actually work (as www.google.com prevents us
from embedding it).

Closes #7