docs(Edge): Document setup for Chromium-based Edge
In a comment from April 2020 [1], the Selenium maintainers said that
they would not support Chromium-based Edge in Selenium 3.  They
instead suggested waiting for a stable version of Selenium 4.
However, as of December 2020, there is no release date in sight for
Selenium 4, Chromium-based Edge is the only version of Edge actively
supported by Microsoft since August 2020, and legacy Edge will reach
end-of-life status in March 2021 [2].

Thankfully, Generic WebDriver Server is generic enough to handle this
case.  We can use it to configure Selenium 3 to respond to Edge and
use Microsoft's `msedgedriver.exe` as a backend executable.

See also https://github.com/SeleniumHQ/selenium/issues/8993

[1]: https://github.com/SeleniumHQ/selenium/issues/8237#issuecomment-629851734
[2]: https://techcommunity.microsoft.com/t5/microsoft-365-blog/microsoft-365-apps-say-farewell-to-internet-explorer-11-and/ba-p/1591666