feat(Xbox): New backend for Xbox One
This backend requires:
 - Xbox One device in Dev Mode
 - Windows 10 host
 - Visual Studio
 - Windows 10 SDK

It builds a UWP (Universal Windows Platform) application hosting a
WebView pointing to the desired destination URL, then deploys that app
to the Xbox and launches it.  In this way, it is very similar to the
Tizen backend.

But because this backend uses Visual Studio to build a UWP app, it
**must** be run on Windows.