docs: Fix reference to non-existent package
There is no NPM package for the "fake" backend.  There would be no
point in publishing it.  But the docs were following the same
structure as every other backend, and gave instructions for
installation from NPM.  This could lead to an attack in which someone
registers the package name and is tricked by our docs into installing
something malicious.

This updates the docs for the "fake" backend to remove any reference
to a package name or installation via NPM.