feat: Expose cast platform APIs to iframe
By forcing an iframe's contents into the same origin as the receiver
app, we gain the ability to expose the cast platform APIs to the
contained iframe.  This won't work necessarily work for every possible
page you could load, so this can be disabled with a command-line flag.

The new receiver code is backward compatible with the v1.0.x version
of these tools.