fix(chromecast): Speed up shutdown (#78)
After we launch the requested receiver app, we should shut down the connection explicitly.  Otherwise, the socket has to time out before the calling application in nodejs can terminate.  This is because nodejs keeps the main thread running so long as anything async could occur, including events on handles and sockets.