feat!(chromecast): Fork receiver app v1
To prepare for a v2 client and receiver app, this copies receiver.html (with some additional comments) to receiver-v1.html.  After merging this change, the registration for the v1 receiver app ID will be changed to point to receiver-v1.html.  This will free us to make breaking changes in receiver.html under a new app ID.