feat(chromecast)!: Fork receiver app v1 (#94)
To prepare for a v2 client and receiver app, this copies receiver.html (with some additional comments) to receiver-v1.html.  After merging this change, the registration for the v1 receiver app ID will be changed to point to receiver-v1.html.  This will free us to make breaking changes in receiver.html under a new app ID.

It is important for a smooth transition to make a copy of receiver.html first and not simply rename it, because there will be a transition period where some Chromecasts under test will still refer to the old URL.  This should be < 48 hours after merge.