feat(chromecast)!: Add v2 receiver app with a redirect mode
This introduces an optional redirect mode, which changes the top URL instead of
using an iframe to host content.  This requires the Cast SDK to be loaded at
the destination URL.  This can be used for Shaka Player testing, but not for
every URL you might want to see on a Chromecast.

This requires deployment to a new receiver app ID.