fix(ci): Fix multi-package release and implement trusted publishing
This does several things:
 - Fixes and updates the multi-package release workflow (#117)
 - Implements trusted publishing (shaka-project/shaka-player#9132)
 - Allows the workflow to work with forks (important for testing)

To work with forks, the workflow has to update a URL from package.json
right before publishing, and to allow that, we have to remove the "clean
git" check from package.json.  That check dates back to when we
published packages manually, so it's safe to remove now that we release
through automated workflows.

Closes #117